- Update postBlockHeight in the database for the posts that were succesfully published on-chain - Separate server logic from logic to prove and update on-chain state
