+++ package.json
@@ -5,7 +5,8 @@
   "main": "main.js",
   "scripts": {
     "build": "tsc -p tsconfig.json",
-    "start": "node ./build/main.js"
+    "service": "node ./build/service.js",
+    "prover": "node ./build/prover.js"
   },
   "keywords": [],
   "author": "chrlyz",
